backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sentence-transformers
faiss-cpu
numpy
scikit-learn
python-multipart
jinja2
requests
openai

prometheus_client
pytest
